Enhance storage functionality with cancellation support and metadata updates + Nuget

+++ Tests/ManagedCode.Storage.Tests/Storages/Abstracts/BlobTests.cs
@@ -51,7 +51,7 @@ public abstract class BlobTests<T> : BaseContainer<T> where T : IContainer
             .BeTrue();
         result.Value!.Length
             .Should()
-            .Be(fileInfo.Length);
+            .Be((ulong)fileInfo.Length);
         result.Value!.Name
             .Should()
             .Be(fileInfo.Name);
